unit tests and Paolo's comment

+++ clients/src/main/java/io/strimzi/common/configuration/http/HttpConsumerConfiguration.java
@@ -6,7 +6,7 @@ package io.strimzi.common.configuration.http;
 
 import java.util.Map;
 
-import static io.strimzi.common.configuration.ClientsConfigurationUtils.parseIntOrDefault;
+import static io.strimzi.common.configuration.ClientsConfigurationUtils.parseLongOrDefault;
 import static io.strimzi.common.configuration.ClientsConfigurationUtils.parseStringOrDefault;
 import static io.strimzi.common.configuration.Constants.CLIENT_ID_ENV;
 import static io.strimzi.common.configuration.Constants.DEFAULT_CLIENT_ID;
@@ -20,15 +20,15 @@ import static io.strimzi.common.configuration.Constants.POLL_TIMEOUT_ENV;
 public class HttpConsumerConfiguration extends HttpClientsConfiguration {
     private final String clientId;
     private final String groupId;
-    private final int pollInterval;
-    private final int pollTimeout;
+    private final long pollInterval;
+    private final long pollTimeout;
 
     public HttpConsumerConfiguration(Map<String, String> map) {
         super(map);
         this.clientId = parseStringOrDefault(map.get(CLIENT_ID_ENV), DEFAULT_CLIENT_ID);
         this.groupId = parseStringOrDefault(map.get(GROUP_ID_ENV), DEFAULT_GROUP_ID);
-        this.pollInterval = parseIntOrDefault(map.get(POLL_INTERVAL_ENV), DEFAULT_POLL_INTERVAL);
-        this.pollTimeout = parseIntOrDefault(map.get(POLL_TIMEOUT_ENV), DEFAULT_POLL_TIMEOUT);
+        this.pollInterval = parseLongOrDefault(map.get(POLL_INTERVAL_ENV), DEFAULT_POLL_INTERVAL);
+        this.pollTimeout = parseLongOrDefault(map.get(POLL_TIMEOUT_ENV), DEFAULT_POLL_TIMEOUT);
     }
 
     public String getClientId() {
@@ -39,11 +39,11 @@ public class HttpConsumerConfiguration extends HttpClientsConfiguration {
         return groupId;
     }
 
-    public int getPollInterval() {
+    public long getPollInterval() {
         return pollInterval;
     }
 
-    public int getPollTimeout() {
+    public long getPollTimeout() {
         return pollTimeout;
     }
